Cache the version data
Maintains a disk and memory cache of the version list.
This should dramatically speed up loading of large queues and when scanning for games since it eliminates most network requests.
It also includes a preloaded cache so it works out of the box and user doesn't have to build the cache themselves.