Fix some error.
CDNSP V5.0.2 + Complete Chinese patch (fixes the bug that the version number is not displayed)
The new version is now ready to update the GUI version and language version directly. However, the new version of the translated file is still a little bit wrong and some operations that can cause obsessive-compulsive disorder. So I re-edited it.